requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py37_0
asn1crypto=1.2.0=py37_0
attrs=19.3.0=pypi_0
backcall=0.1.0=py37_0
beautifulsoup4=4.8.1=py37_0
blas=1.0=mkl
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37hb5b8e2f_0
chardet=3.0.4=py37_1003
click=7.0=pypi_0
colorama=0.4.1=pypi_0
cryptography=2.8=py37ha12b0ac_0
cycler=0.10.0=pypi_0
decorator=4.4.1=py_0
feedparser=5.2.1=py37_1
filelock=3.0.12=pypi_0
fpdf=1.7.2=pypi_0
freetype=2.10.0=h24853df_1
funcsigs=1.0.2=pypi_0
icu=58.2=h4b95b61_1
idna=2.8=py37_0
importlib-metadata=0.23=pypi_0
intel-openmp=2019.4=233
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.9.0=py37h39e3cac_0
ipython-sql=0.3.9=py37_1000
ipython_genutils=0.2.0=py37_0
jedi=0.15.1=py37_0
jpeg=9c=h1de35cc_1001
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.0=py37_0
kaggle=1.5.6=pypi_0
kiwisolver=1.1.0=pypi_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.37=h2573ce8_0
libsodium=1.0.16=h3efe00b_0
libtiff=4.0.10=hd08fb8f_1003
libxml2=2.9.9=hf6e021a_1
libxslt=1.1.33=h33a18ac_0
lxml=4.4.2=pypi_0
lz4-c=1.8.3=h6de7cb9_1001
matplotlib=3.1.2=pypi_0
missingno=0.4.2=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.14=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
modin=0.6.3=pypi_0
more-itertools=7.2.0=pypi_0
ncurses=6.1=h0a44026_1
numpy=1.17.2=py37h99e6662_0
numpy-base=1.17.2=py37h6575580_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1d=h0b31af3_0
packaging=19.2=pypi_0
pandas=0.25.3=py37h0a44026_0
parso=0.5.1=py_0
patsy=0.5.1=py37_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.0=py37hb68e598_0
pip=19.3.1=py37_0
pluggy=0.13.1=pypi_0
prettytable=0.7.2=py_3
prompt_toolkit=2.0.10=py_0
protobuf=3.11.0=pypi_0
psutil=5.6.7=pypi_0
ptyprocess=0.6.0=py37_0
py=1.8.0=pypi_0
pycparser=2.19=py37_0
pygments=2.4.2=py_0
pymysql=0.9.3=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.5=pypi_0
pysocks=1.7.1=py37_0
pytest=5.3.1=pypi_0
python=3.7.5=h359304d_0
python-dateutil=2.8.0=py37_0
python-pptx=0.6.18=py_0
python-slugify=4.0.0=pypi_0
pytz=2019.3=py_0
pyyaml=5.1.2=pypi_0
pyzmq=18.1.0=py37h0a44026_0
ray=0.7.3=pypi_0
readline=7.0=h1de35cc_5
redis=3.3.11=pypi_0
requests=2.22.0=py37_1
requests-oauthlib=1.3.0=pypi_0
scipy=1.3.1=py37h1410ff5_0
seaborn=0.9.0=pypi_0
setuptools=41.6.0=py37_0
six=1.12.0=py37_0
soupsieve=1.9.5=py37_0
sqlalchemy=1.3.11=py37h1de35cc_0
sqlite=3.30.1=ha441bb4_0
sqlparse=0.3.0=py_0
statsmodels=0.10.1=py37h1d22016_0
text-unidecode=1.3=pypi_0
tk=8.6.8=ha441bb4_0
tornado=6.0.3=py37h1de35cc_0
tqdm=4.41.0=pypi_0
traitlets=4.3.3=py37_0
tweepy=3.8.0=pypi_0
urllib3=1.24.3=pypi_0
wcwidth=0.1.7=py37_0
wget=3.2=pypi_0
wheel=0.33.6=py37_0
xlrd=1.2.0=pypi_0
xlsxwriter=1.2.7=py_0
xz=5.2.4=h1de35cc_4
zeromq=4.3.1=h0a44026_3
zipp=0.6.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.0=ha9f0a20_0
